requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bert==2.2.0
bleach==1.5.0
boto3==1.12.30
botocore==1.15.30
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
erlastic==2.0.0
filelock==3.0.12
gast>=0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.5.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras>=2.3.1
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-pretrained-bert==0.4.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
seqeval==0.0.12
six==1.14.0
tensorboard>=2.1.1
tensorboard-plugin-wit>=1.6.0.post2
tensorflow>=2.1.0
tensorflow-estimator>=2.1.0
tensorflow-tensorboard>=1.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tokenizers==0.5.2
torch==1.4.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
transformers==2.6.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
